// src/orders/interfaces/order-products.interface.ts
import { OrderStatus } from '@prisma/client';

export interface OrderProducts {
  OrderItem: {
    name: string;
    productId: number;
    quantity: number;
    price: number;
  }[];
  id: string;
  totalAmount: number;
  totalItems: number;
  status: OrderStatus;
  paid: boolean;
  paidAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
}